[update] 데미지 표시 빅인티져화 작업
추후 픽셀몬 기본스탯도 빅인티저화 완성 시 실질적 가능

+++ Assets/02.Scripts/Entity/HealthSystem/BossHealthSystem.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using System.Numerics;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
@@ -9,8 +10,8 @@ public class BossHealthSystem : MonoBehaviour
     [SerializeField] private Slider bossHpBar;
     [SerializeField] private TextMeshProUGUI bossHpTxt;
 
-    [SerializeField] private float currentHealth => enemy.healthSystem.currentHealth;
-    [SerializeField] private float maxHealth => enemy.healthSystem.maxHealth;
+    [SerializeField] private BigInteger currentHealth => enemy.healthSystem.currentHealth;
+    [SerializeField] private BigInteger maxHealth => enemy.healthSystem.maxHealth;
 
     Coroutine bossCoroutine;
 
@@ -25,7 +26,7 @@ public class BossHealthSystem : MonoBehaviour
     {
         while(true)
         {
-            bossHpBar.value = currentHealth / maxHealth;
+            bossHpBar.value = (float)(currentHealth / maxHealth);
             bossHpTxt.text = ((int)(bossHpBar.value * 100)).ToString() + "%";
 
             if (currentHealth <= 0)
